Fixes viktorstrate/photoview#71 - Performance optimizations - restructured the code so no necessary Layout and AlbumFilter component rerenderings are made.
